+++ RELEASENOTES.md
@@ -2,6 +2,12 @@
 
 ### Next version
 
+### v3.0.3 2024-03-23
+- eGMP: detection of AC/DC based on power kw
+- contributing data: added battery current amps (for abrp support)
+- gps altitude fix
+- speed screen: outdoor temperature instead inverter temperature
+
 ### v3.0.2 2024-03-20
 - Improved logic for stop CAN queue - not only car off but not charging too (works with DC charger)
 - Average charging speed in kW